Corrigindo a mensagem assustadora e melhorando algumas mensagens de erro
