+++ src/br/univali/portugol/nucleo/execucao/erros/ErroValorVetorMatrizNaoInicializado.java
@@ -1,11 +1,15 @@
 package br.univali.portugol.nucleo.execucao.erros;
 
+import br.univali.portugol.nucleo.GeradorDeExemplosDeInicializacao;
 import br.univali.portugol.nucleo.asa.ExcecaoVisitaASA;
 import br.univali.portugol.nucleo.asa.NoReferencia;
 import br.univali.portugol.nucleo.asa.NoReferenciaMatriz;
 import br.univali.portugol.nucleo.asa.NoReferenciaVetor;
 import br.univali.portugol.nucleo.asa.VisitanteASABasico;
 import br.univali.portugol.nucleo.mensagens.ErroExecucao;
+import br.univali.portugol.nucleo.simbolos.Matriz;
+import br.univali.portugol.nucleo.simbolos.Simbolo;
+import br.univali.portugol.nucleo.simbolos.Vetor;
 
 /**
  *
@@ -14,20 +18,23 @@ import br.univali.portugol.nucleo.mensagens.ErroExecucao;
 public final class ErroValorVetorMatrizNaoInicializado extends ErroExecucao
 {
     private final NoReferencia referencia;
+    private final Simbolo simbolo;
     private int linha;
     private int coluna;
     private int indice;
-    
-    public ErroValorVetorMatrizNaoInicializado(NoReferenciaVetor referenciaVetor, int indice)
+
+    public ErroValorVetorMatrizNaoInicializado(Vetor vetor, NoReferenciaVetor referenciaVetor, int indice)
     {
+        this.simbolo = vetor;
         this.referencia = referenciaVetor;
         this.indice = indice;
         this.setLinha(referencia.getTrechoCodigoFonteNome().getLinha());
-        this.setColuna(referencia.getTrechoCodigoFonteNome().getColuna());        
+        this.setColuna(referencia.getTrechoCodigoFonteNome().getColuna());
     }
 
-    public ErroValorVetorMatrizNaoInicializado(NoReferenciaMatriz referenciaMatriz, int linha, int coluna)
+    public ErroValorVetorMatrizNaoInicializado(Matriz matriz, NoReferenciaMatriz referenciaMatriz, int linha, int coluna)
     {
+        this.simbolo = matriz;
         this.referencia = referenciaMatriz;
         this.linha = linha;
         this.coluna = coluna;
@@ -37,22 +44,22 @@ public final class ErroValorVetorMatrizNaoInicializado extends ErroExecucao
 
     /**
      * {@inheritDoc }
-     * 
-     * @return  o texto da mensagem
+     *
+     * @return o texto da mensagem
      */
     @Override
     protected String construirMensagem()
-    {        
+    {
         return new ConstrutorMensagem().construirMensagem();
     }
-    
+
     private final class ConstrutorMensagem extends VisitanteASABasico
     {
         public ConstrutorMensagem()
         {
-            
-        }        
-        
+
+        }
+
         public String construirMensagem()
         {
             try
@@ -63,18 +70,58 @@ public final class ErroValorVetorMatrizNaoInicializado extends ErroExecucao
             {
                 return e.getMessage();
             }
-        } 
+        }
 
         @Override
         public Object visitar(NoReferenciaVetor noReferenciaVetor) throws ExcecaoVisitaASA
         {
-            return String.format("O elemento do vetor '%s', na posição [%d] não foi inicializado", noReferenciaVetor.getNome(),indice);
+            StringBuilder construtorTexto = new StringBuilder();
+
+            construtorTexto.append(String.format("O elemento do vetor '%s', na posição [%d] não foi inicializado", noReferenciaVetor.getNome(), indice));
+            construtorTexto.append(". Você pode inicializar o elemento atribuindo um valor do tipo \"");
+            construtorTexto.append(simbolo.getTipoDado().getNome());
+            construtorTexto.append("\". Exemplo: ");
+            construtorTexto.append(simbolo.getNome());
+            construtorTexto.append("[");
+            construtorTexto.append(indice);
+            construtorTexto.append("] = ");
+            construtorTexto.append(GeradorDeExemplosDeInicializacao.gerarValorAleatorio(simbolo.getTipoDado()));
+            construtorTexto.append(". Você também pode usar a função \"leia\" para ler um valor digitado pelo usuário. Exemplo: ");
+            construtorTexto.append("leia(");
+            construtorTexto.append(simbolo.getNome());
+            construtorTexto.append("[");
+            construtorTexto.append(indice);
+            construtorTexto.append("])");
+
+            return construtorTexto.toString();
         }
 
         @Override
         public Object visitar(NoReferenciaMatriz noReferenciaMatriz) throws ExcecaoVisitaASA
         {
-            return String.format("O elemento da matriz '%s', na linha [%d] e coluna [%d] não foi inicializado", noReferenciaMatriz.getNome(), linha, coluna);
+            StringBuilder construtorTexto = new StringBuilder();
+
+            construtorTexto.append(String.format("O elemento da matriz '%s', na linha [%d] e coluna [%d] não foi inicializado", noReferenciaMatriz.getNome(), linha, coluna));
+            construtorTexto.append(". Você pode inicializar o elemento atribuindo um valor do tipo \"");
+            construtorTexto.append(simbolo.getTipoDado().getNome());
+            construtorTexto.append("\". Exemplo: ");
+            construtorTexto.append(simbolo.getNome());
+            construtorTexto.append("[");
+            construtorTexto.append(linha);
+            construtorTexto.append("][");
+            construtorTexto.append(coluna);
+            construtorTexto.append("] = ");
+            construtorTexto.append(GeradorDeExemplosDeInicializacao.gerarValorAleatorio(simbolo.getTipoDado()));
+            construtorTexto.append(". Você também pode usar a função \"leia\" para ler um valor digitado pelo usuário. Exemplo: ");
+            construtorTexto.append("leia(");
+            construtorTexto.append(simbolo.getNome());
+            construtorTexto.append("[");
+            construtorTexto.append(linha);
+            construtorTexto.append("][");
+            construtorTexto.append(coluna);
+            construtorTexto.append("])");
+
+            return construtorTexto.toString();
         }
     }
 }
